Use individual files to store related types and interfaces
