+++ src/index.ts
@@ -1,7 +1,7 @@
 import ExceptionLogger from './ExceptionLogger';
 import Logger from './Logger';
 import LogReporter from './LogReporter';
-import { LogType } from './types';
+import { LogType } from './types/logger';
 
 export {
   ExceptionLogger,
