+++ src/types/general.ts
@@ -1,0 +1,10 @@
+import { SlackMessageSection } from './slack';
+
+export interface LoggerOptions {
+    group?: any,
+    slackMessageExtraSections?: SlackMessageSection[],
+};
+
+export interface GeneralOptions {
+    slackIntegration?: boolean,
+}
