+++ src/types/logger.ts
@@ -1,0 +1,8 @@
+export enum LogType {
+    TRACE = 'TRACE',
+    DEBUG = 'DEBUG',
+    INFO = 'INFO',
+    WARN = 'WARN',
+    ERROR = 'ERROR',
+    FATAL = 'FATAL',
+}
