+++ src/types/slack.ts
@@ -1,0 +1,9 @@
+export interface SlackMessageSectionField {
+    type: string,
+    text: string,
+}
+
+export interface SlackMessageSection {
+    type: string,
+    fields: SlackMessageSectionField[]
+}
